6047: added feature mssqlserver / tsql adapter support
Implemented support for sql files from MS SQL Server.
The manifest.json containes the adapter name 'sqlserver' if the dbt core project
has been initialized with the ms sql server adapter.
The sqlglot sql parser utilizes the term 'tsql' for this sql dialect.
Supporting this dialect is basically a one-liner to match 'sqlserver' to
'tsql'.

The implementation provides two means to provide an adapter match.
* command line argument (CLI) --adapter <adapter-name>
	this option overrides everything that is dynamically parsed /
	matched from manifest.json

* matching class that contains a simple dictionary mapping from one name
  to another, that could be used for future conflicts with adapter names
  see registry.py

Instead of using the adapter name from the manifest.json file directly,
the registry is used to potentially map different adapter names to
sqlglot names. The CLI argument overrides any of this mechanism and
enforces the adapter name.